Bring back attachments, greatly simplified.
